+++ src/what/inbox/InboxActivity.java
@@ -126,7 +126,7 @@ public class InboxActivity extends MyActivity implements OnClickListener {
 	private void openMessage(int i) {
 		Bundle b = new Bundle();
 		intent = new Intent(InboxActivity.this, what.forum.ThreadActivity.class);
-		b.putInt("id", inbox.getResponse().getMessages().get(i).getConvId());
+		b.putInt("id", inbox.getResponse().getMessages().get(i).getConvId().intValue());
 		intent.putExtras(b);
 		startActivityForResult(intent, 0);
 	}
